pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "northstar_bq"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = []  # core stays lean

[project.optional-dependencies]
bigquery = [
	"google-cloud-bigquery>=3.25.0",
	"google-auth>=2.33.0",
]

[tool.setuptools]
packages = ["core", "experts", "verify", "src", "retrieval", "bq", "pipeline"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
